refactor: deduplicate shared code across SwiftUI components
Extract duplicated code into shared internal helpers, reducing ~470 lines
of duplication across the codebase.

**Shared ButtonStyle:**
- Extract `LemonadePressTrackingButtonStyle` from 4 identical private
  copies in Button, SegmentedControl, IconButton, and Link

**Shared TextField helpers:**
- Extract `TextFieldConstants`, `TextFieldLabelRow`, `TextFieldSupportText`,
  `TextFieldContainerModifier`, and color computation functions into
  `LemonadeTextFieldShared.swift`
- TextField reduced from 1310 to ~970 lines
- SelectField reduced from 264 to ~196 lines
- Eliminates ~400 lines of copy-pasted code (backgroundColor, borderColor,
  label row, support text repeated 4-5 times)

**Shared utilities:**
- Extract `applyIf` View extension to `View+Lemonade.swift`
- Extract `StatefulPreviewWrapper` to `LemonadePreviewHelpers.swift`

**Centralized font family:**
- Add `LemonadeTypography.fontFamily` constant ("Figtree")
- Replace hardcoded "Figtree" strings in Text, Badge, and Chip
- Add doc comment explaining bodyMediumBold = semibold (Figtree limitation)

**Regenerated Xcode project** via XcodeGen to include new files.
